+++ demisto_sdk/commands/common/content/objects/pack_objects/indicator_field/indicator_field.py
@@ -4,7 +4,7 @@ from typing import Union
 
 import demisto_client
 from demisto_sdk.commands.common.constants import (INCIDENT_FIELD,
-                                                   INDICATOR_FIELD)
+                                                   INDICATOR_FIELD, FileType)
 from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.json_content_object import \
     JSONContentObject
 from wcmatch.pathlib import Path
@@ -60,3 +60,6 @@ class IndicatorField(JSONContentObject):
             indicator_fields_unified_file.write(bytes(json.dumps(indicator_fields_unified_data), 'utf-8'))
             indicator_fields_unified_file.seek(0)
             return client.import_incident_fields(file=indicator_fields_unified_file.name)
+
+    def type(self):
+        return FileType.INDICATOR_FIELD