+++ demisto_sdk/commands/common/tools.py
@@ -363,7 +363,7 @@ def get_file(method, file_path, type_of_file):
                 print_error(
                     "{} has a structure issue of file type{}. Error was: {}".format(file_path, type_of_file, str(e)))
                 return {}
-    if type(data_dictionary) is dict:
+    if isinstance(data_dictionary, (dict, list)):
         return data_dictionary
     return {}
 
@@ -895,7 +895,7 @@ def find_type(path: str = '', _dict=None, file_type: Optional[str] = None, ignor
         return FileType.CHANGELOG
 
     # integration image
-    if path.endswith('_image.png'):
+    if path.endswith('_image.png') and not path.endswith("Author_image.png"):
         return FileType.IMAGE
 
     # doc files images